+++ crates/driver/src/infra/solver/dto/solution.rs
@@ -37,12 +37,7 @@ impl Solutions {
                                         "invalid order UID specified in fulfillment"
                                     ))?
                                     .clone();
-
-                                let uniform_sell_price =
-                                    solution.prices[&order.sell.token.wrap(weth).into()];
-                                let uniform_buy_price =
-                                    solution.prices[&order.buy.token.wrap(weth).into()];
-
+                                
                                 competition::solution::trade::Fulfillment::new(
                                     order,
                                     fulfillment.executed_amount.into(),
@@ -52,8 +47,6 @@ impl Solutions {
                                         ),
                                         None => competition::solution::trade::Fee::Static,
                                     },
-                                    uniform_sell_price,
-                                    uniform_buy_price,
                                 )
                                 .map(competition::solution::Trade::Fulfillment)
                                 .map_err(
